Include chart area type
