+++ .eslintrc.yml
@@ -5,6 +5,7 @@ extends: [
 globals:
   require: true
   config: true
+  window: true
   __DEV__: true
 parser: babel-eslint
 parserOptions:
